test(segment): migrate events
